Adopt SciMLTesting v1.2 folder-based run_tests (on top of grouped-tests conversion)
MTK's root test suite is a capability-based monorepo: the named groups
(InterfaceI/II, Initialization, SymbolicIndexingInterface, Extensions,
Optimization, Downstream, FMI) draw test files from the SAME shared
ModelingToolkitBase/test directory (via the @mtktestset macro) partitioned by
group membership, plus a handful of root test/*.jl files; they do NOT partition
into per-group folders and there is no Core group. So this uses v1.2's
explicit-args run_tests(groups=..., qa=..., all=...) rather than folder
discovery, as the harness recommends for shared-module suites.

Each group's body is extracted verbatim into a per-group entry file
(test/group_<name>.jl) that includes test/shared/mtktestset.jl (the @mtktestset
macro + MTKBasePath) and runs the group's original @safetestset/@mtktestset
lines unchanged — same tests, same isolation. The per-group activate_*_env
helpers are replaced by run_tests env=/parent= (parent = [MTKBasePath, repo
root], matching the old develop set); QA -> run_qa with the existing
target_defined_modules JET kwargs. Curated `all` lists all 9 groups in the
original order so GROUP=All is preserved. Adds SciMLTesting (+SafeTestsets) to
the test target and every @safetestset sub-env (qa, downstream, fmi,
MTKBase/test/extensions, MTKBase/test/optimization); drops Pkg from the root
test deps and the qa env (only the old activate harness used it).
test_groups.toml unchanged. Behavior-preserving.
