+++ test/group_fmi.jl
@@ -1,0 +1,3 @@
+include("shared/mtktestset.jl")
+
+@safetestset "FMI Extension Test" include("fmi/fmi.jl")
