+++ test/group_symbolicindexinginterface.jl
@@ -1,0 +1,4 @@
+include("shared/mtktestset.jl")
+
+@mtktestset("SciML Problem Input Test", "sciml_problem_inputs.jl")
+@mtktestset("MTKParameters Test", "mtkparameters.jl")
