+++ test/qa/qa.jl
@@ -1,12 +1,12 @@
 using ModelingToolkit
 using Aqua
 using JET
-using Test
+using SciMLTesting
 
-@testset "Aqua" begin
-    Aqua.test_all(ModelingToolkit)
-end
-
-@testset "JET" begin
-    JET.test_package(ModelingToolkit; target_defined_modules = true)
-end
+run_qa(
+    ModelingToolkit;
+    Aqua = Aqua,
+    JET = JET,
+    jet = true,
+    jet_kwargs = (; target_defined_modules = true),
+)
